sprout-backend/requirements/base.txt
Django==1.6
South==0.8.4
dj-static==0.0.5
django-cors-headers==0.12
djangorestframework==2.3.10
longerusername==0.4
pystache==0.5.3
static==1.0.2
wsgiref==0.1.2
